mcp-template/requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.9
pymysql>=1.1.0
pyodbc>=5.1.0
